LayoutTests/platform/chromium-mac/fast/js/deep-recursion-test-expected.txt

This test how deep we can recurse, and that we get an exception when we do, as opposed to a stack overflow.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL msg should be RangeError: Maximum call stack size exceeded.. Was RangeError: Maximum call stack size exceeded.
FAIL msg should be RangeError: Maximum call stack size exceeded.. Was RangeError: Maximum call stack size exceeded.
FAIL msg should be RangeError: Maximum call stack size exceeded.. Was RangeError: Maximum call stack size exceeded.
FAIL msg should be RangeError: Maximum call stack size exceeded.. Was RangeError: Maximum call stack size exceeded.
PASS successfullyParsed is true

TEST COMPLETE
